+++ src/main/java/com/contractor/enums/UserRoleProperty.java
@@ -1,0 +1,7 @@
+package com.contractor.enums;
+
+public enum UserRoleProperty {
+    administrator,
+    contractor,
+    customer,
+}