+++ src/main/java/com/contractor/model/dao/JobCategoryDao.java
@@ -1,0 +1,9 @@
+package com.contractor.model.dao;
+
+import com.contractor.model.entity.JobCategory;
+
+public class JobCategoryDao extends AbstractDao<JobCategory> {
+    public JobCategoryDao() {
+        super(JobCategory.class);
+    }
+}