+++ src/main/java/com/contractor/model/dao/JobDao.java
@@ -1,7 +1,6 @@
 package com.contractor.model.dao;
 
 import com.contractor.model.entity.Job;
-import com.contractor.model.entity.JobType;
 
 public class JobDao extends AbstractDao<Job> {
     public JobDao() {