+++ src/main/java/com/contractor/model/dao/JobTypeDao.java
@@ -1,9 +1,0 @@
-package com.contractor.model.dao;
-
-import com.contractor.model.entity.JobType;
-
-public class JobTypeDao extends AbstractDao<JobType> {
-    public JobTypeDao() {
-        super(JobType.class);
-    }
-}